Redesigned the hierarchy of modules in the project
