+++ body/__init__.py
@@ -1,0 +1,2 @@
+from .py_widjet_alarm_clock import PyAlarmClock
+from .main import Ui_Body
