+++ body/body_style.css
@@ -1,0 +1,11 @@
+@import url("py_widjet_alarm_clock/alarm_clock_style.css");
+
+.alarm_clock_area {
+    background: none;
+    border: none;
+}
+
+.alarm_clock_scroll_area {
+    border: none;
+    background-color: rgba(0, 0, 0, 0);
+}
