+++ body/py_widjet_alarm_clock/__init__.py
@@ -1,1 +1,2 @@
+from .py_toggle import PyToggle
 from .py_alarm_clock import PyAlarmClock
