+++ body/py_widjet_alarm_clock/alarm_clock_style.css
@@ -1,0 +1,8 @@
+.alarm_clock {
+    background: #fff5ff;
+    border-radius: 50%;
+}
+
+.space_for_time {
+    border: none;
+}
