+++ header/header_style.css
@@ -1,0 +1,12 @@
+.circle_bg {
+    background-color: #282a36;
+    border-radius: 310px;
+}
+
+.header {
+    background: none;
+}
+
+.time_label {
+    color: #f8f8f2;
+}
