+++ header/main.py
@@ -1,0 +1,25 @@
+from PyQt6.QtCore import *
+from PyQt6.QtGui import QFont
+from PyQt6.QtWidgets import *
+
+
+class Ui_Header:
+    def __init__(self):
+        self.circle_bg = QFrame()
+        self.circle_bg.setGeometry(QRect(0, 0, 620, 620))
+        self.circle_bg.setProperty("class", "circle_bg")
+
+        # application header, with time and additional information
+        self.header = QFrame(self.circle_bg)
+        self.header.setGeometry(QRect(0, 0, 620, 230))
+        self.header.setProperty("class", "header")
+
+        self.font_time = QFont()
+        self.font_time.setFamily("Segoe UI")
+        self.font_time.setPointSize(75)
+
+        self.time_label = QLabel("00:00", self.header)
+        self.time_label.setGeometry(QRect(0, 0, 620, 230))
+        self.time_label.setAlignment(Qt.AlignmentFlag.AlignHCenter | Qt.AlignmentFlag.AlignVCenter)
+        self.time_label.setFont(self.font_time)
+        self.time_label.setProperty("class", "time_label")
